Update (un-update?) world implementation to b1.7.3

+++ TrueCraft.API/World/IChunk.cs
@@ -10,11 +10,11 @@ namespace TrueCraft.API.World
         ISection[] Sections { get; }
         byte[] Biomes { get; }
         DateTime LastAccessed { get; set; }
-        short GetBlockID(Coordinates3D coordinates);
+        byte GetBlockID(Coordinates3D coordinates);
         byte GetMetadata(Coordinates3D coordinates);
         byte GetSkyLight(Coordinates3D coordinates);
         byte GetBlockLight(Coordinates3D coordinates);
-        void SetBlockID(Coordinates3D coordinates, short value);
+        void SetBlockID(Coordinates3D coordinates, byte value);
         void SetMetadata(Coordinates3D coordinates, byte value);
         void SetSkyLight(Coordinates3D coordinates, byte value);
         void SetBlockLight(Coordinates3D coordinates, byte value);
